pyproject.toml
[build-system]
requires = ["setuptools>=64.0.0", "wheel"]
build-backend = "setuptools.build_meta:__legacy__"

[project]
name = "snc"
version = "0.1.0"
description = "A system narrative compiler for allowing developers to write code in natural, narrative language."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "fastapi[all]>=0.100.0,<0.101.0",
    "psycopg2-binary>=2.9.5,<2.10.0",
    "sqlalchemy>=2.0.4,<2.1.0",
    "alembic>=1.10.2,<1.11.0",
    "python-dotenv>=1.0.0,<1.1.0",
    "pydantic>=1.8.0",
    "pydantic-settings>=2.0.0",
    "openai>=1.57.4,<1.58.0",
    "groq>=0.13.1,<0.14.0",
    "pyjwt>=2.6.0,<2.7.0",
    "requests>=2.31.0,<2.32.0",
    "lark-parser>=0.12.0,<0.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0,<8.2.0",
    "pytest-benchmark>=5.1.0,<5.2.0",
    "pytest-cov>=4.1.0,<4.2.0",
    "black>=23.10.1,<23.11.0",
    "flake8>=4.0.0",
    "flake8-docstrings>=1.7.0",
    "mypy>=0.950",
    "pre-commit>=3.5.0",
    "honcho>=1.1.0,<1.2.0",
]
test = [
    "pytest>=8.1.0,<8.2.0",
    "pytest-cov>=4.1.0,<4.2.0",
    "pytest-benchmark>=5.1.0,<5.2.0",
]
docs = [
    "sphinx>=4.5.0",
    "sphinx-rtd-theme>=1.0.0",
    "sphinx-autodoc-typehints>=1.18.0",
    "sphinx-copybutton>=0.5.0",
    "myst-parser>=0.18.0",
]

[project.urls]
"Homepage" = "https://github.com/westonvogelsong/snc-v3"

[project.scripts]
snc = "snc.main:main"
snc-db = "snc.script.create_database:main"

[tool.setuptools]
packages = ["snc"] 